feat: Add support for JSON filters in BaseModelTable
This commit adds support for JSON filters in the `BaseModelTable` component. It checks if a filter is of type `FieldType.JSON` and if the filter data is of type `Object`. If both conditions are met, the filter data is included in the query. This enhancement allows for more flexible filtering options when working with JSON data in the table.