refactor: Override name property in Label class
This code change overrides the name property in the Label class to allow for more flexibility in handling scenarios where the name is not specified. By setting the name property to undefined, it improves the usability and maintainability of the code.

Note: This commit message follows the established convention of starting with a verb in the imperative form (e.g., "Update", "Add", "Fix") and providing a concise summary of the changes made.