refactor: Add checkCreateBlockPermissions method to CreatePermission
This code change adds the checkCreateBlockPermissions method to the CreatePermission class in CreatePermission.ts. The method is responsible for checking the block permissions for creating a new record. It allows the query to run if the system is making the query or if the user has root or master admin privileges. This enhancement provides better control over the creation of records and improves the security of the application.